pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dsiBuildSoftSummative"
authors = [
    {name = "Ka Ho (Gerald) Chan"},
]
description = "Building Robust Software Summative Assignment"
version = "0.0.5"
dependencies = ["matplotlib", "numpy", "pyyaml", "requests"]
requires-python = ">=3.10.0"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
]